Enable modular notifications
Restructure the Discord notified as an implementation of a common Notification System which can be extended
