+++ src/config.rs
@@ -59,12 +59,12 @@ fn default_ignore_older_than() -> std::time::Duration {
 #[derive(Clone, TS, Serialize, Deserialize, Debug)]
 #[ts(export)]
 pub struct NotifierConfig {
-    pub discord: Option<NotifierDiscordConfig>,
+    pub discord: Option<DiscordConfig>,
 }
 
 #[derive(Clone, TS, Serialize, Deserialize, Debug)]
 #[ts(export)]
-pub struct NotifierDiscordConfig {
+pub struct DiscordConfig {
     pub webhook_url: String,
     pub notify_on: Vec<TaskStatus>,
 }
