+++ src/main.rs
@@ -120,7 +120,7 @@ async fn main() -> Result<()> {
     let config = Arc::new(RwLock::new(config));
     let h_scraper = run_module!(bus, module::scraper::RSS::new(config.clone()));
     let h_recorder = run_module!(bus, module::recorder::YTArchive::new(config.clone()));
-    let h_notifier = run_module!(bus, module::notifier::Discord::new(config.clone()));
+    let h_notifier = run_module!(bus, module::notifier::NotificationSystem::new(config.clone()));
     let h_webserver = run_module!(bus, module::web::WebServer::new(config.clone()));
 
     // Listen for signals
